Refactor: Delete FormConatiner.js
